+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All Notable changes to `digipolisgent/toerismevlaanderen-lodging` package.
 
-## [Unreleased]
+## [0.2.1]
 
 ### Fixed
 
@@ -36,6 +36,7 @@ Initial package.
 * Added getting a list of lodgings by given filters.
 * Added getting the details of a single lodging.
 
+[0.2.1]: https://github.com/digipolisgent/php_package_dg-toerismevlaanderen-lodging/compare/0.2.0...0.2.1
 [0.2.0]: https://github.com/digipolisgent/php_package_dg-toerismevlaanderen-lodging/compare/0.1.0...0.2.0
 [0.1.0]: https://github.com/digipolisgent/php_package_dg-toerismevlaanderen-lodging/releases/tag/0.1.0
 [Unreleased]: https://github.com/digipolisgent/php_package_dg-toerismevlaanderen-lodging/compare/master...develop